requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
decouple==0.0.7
Django==5.0.6
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.3.0
requests==2.32.3
sqlparse==0.5.0
Unidecode==1.3.8
urllib3==2.2.1
